Custom Slot: Refactor into functional component
