+++ docs/frontend.rst
@@ -159,15 +159,15 @@ General Components
 +===============================+==================================================+==========================+
 |``alert.jsx``                  | .. image:: components/alert.png                  |                          |
 +-------------------------------+--------------------------------------------------+--------------------------+
-|``calendar.jsx``               | .. image:: components/calendar.png               |                          |
+|``calendar.tsx``               | .. image:: components/calendar.png               |                          |
 +-------------------------------+--------------------------------------------------+--------------------------+
-|``calendar.jsx``               | .. image:: components/calendar.png               |                          |
+|``calendar.tsx``               | .. image:: components/calendar.png               |                          |
 +-------------------------------+--------------------------------------------------+--------------------------+
 |``course_modal_section.jsx``   | .. image:: components/course_modal_section.png   |                          |
 +-------------------------------+--------------------------------------------------+--------------------------+
 |``credit_ticker.jsx``          | .. image:: components/credit_ticker.png          |                          |
 +-------------------------------+--------------------------------------------------+--------------------------+
-|``custom_slot.jsx``            | .. image:: components/custom_slot.png            |                          |
+|``custom_slot.tsx``            | .. image:: components/custom_slot.png            |                          |
 +-------------------------------+--------------------------------------------------+--------------------------+
 |``day_calendar.jsx``           | .. image:: components/day_calendar.png           |                          |
 +-------------------------------+--------------------------------------------------+--------------------------+
